Changed location of navBar tests as placing them in App.test.tsx was breaking the branch protection tests. Created a new file under src/tests/NavBar.test.tsx
